requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
mcp[cli]>=1.3.0
# Google platforms
google-ads>=24.0.0
google-analytics-data>=0.18.0
google-analytics-admin>=0.22.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
# Meta
requests>=2.31.0
# HTTP
httpx>=0.27.0
# Auth / crypto
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# Config
python-dotenv>=1.0.0
# Rate limiting
slowapi>=0.1.9
# Email
resend>=2.0.0
